+++ src/Supervertaler.Trados/Controls/AiAssistantControl.cs
@@ -5,6 +5,7 @@ using System.Drawing.Imaging;
 using System.IO;
 using System.Windows.Forms;
 using Sdl.Desktop.IntegrationApi.Interfaces;
+using Supervertaler.Trados.Core;
 using Supervertaler.Trados.Models;
 
 namespace Supervertaler.Trados.Controls
@@ -142,7 +143,7 @@ namespace Supervertaler.Trados.Controls
             };
             _btnHelp.FlatAppearance.BorderSize = 0;
             _btnHelp.FlatAppearance.MouseOverBackColor = Color.FromArgb(220, 220, 220);
-            _btnHelp.Click += OnHelpClick;
+            _btnHelp.Click += OnHelpDropdown;
 
             Controls.Add(_btnSettings);
             Controls.Add(_btnHelp);
@@ -434,12 +435,38 @@ namespace Supervertaler.Trados.Controls
             _btnSettings.Location = new Point(_btnHelp.Left - _btnSettings.Width, 1);
         }
 
-        private void OnHelpClick(object sender, EventArgs e)
+        private void OnHelpDropdown(object sender, EventArgs e)
         {
-            using (var dlg = new AboutDialog())
+            string topic = _tabControl.SelectedIndex == 1
+                ? HelpSystem.Topics.BatchTranslate
+                : HelpSystem.Topics.AiAssistantChat;
+            string label = _tabControl.SelectedIndex == 1
+                ? "Batch Translate Help"
+                : "Assistant Help";
+
+            var menu = new ContextMenuStrip();
+            menu.Items.Add(label, null, (s, ev) =>
+                HelpSystem.OpenHelp(topic));
+            menu.Items.Add("-");  // separator
+            menu.Items.Add("About Supervertaler for Trados", null, (s, ev) =>
             {
-                dlg.ShowDialog(FindForm());
+                using (var dlg = new AboutDialog())
+                    dlg.ShowDialog(FindForm());
+            });
+            menu.Show(_btnHelp, new Point(0, _btnHelp.Height));
+        }
+
+        protected override bool ProcessCmdKey(ref Message msg, Keys keyData)
+        {
+            if (keyData == Keys.F1)
+            {
+                string topic = _tabControl.SelectedIndex == 1
+                    ? HelpSystem.Topics.BatchTranslate
+                    : HelpSystem.Topics.AiAssistantChat;
+                HelpSystem.OpenHelp(topic);
+                return true;
             }
+            return base.ProcessCmdKey(ref msg, keyData);
         }
 
         private void LayoutInputPanel()